+++ Code/Libraries/Foundation/Platform/Windows/WString.cpp
@@ -19,8 +19,7 @@ WString::WString(const wchar_t* str) : mSize(0)
 }
 
 WString::WString(const wchar_t* str, size_t lengthInWChars) :
-    // +1 for the null terminator (unlike String, WString actually adds to the
-    // size for null)
+    // +1 for the null terminator (unlike String, WString actually adds to the size for null)
     mSize(lengthInWChars + 1)
 {
   if (!str)
@@ -40,12 +39,12 @@ WString::WString(StringParam str) : mSize(0)
   if (str.Empty())
     return;
   // first read the number of wide characters we need to allocate for our
-  // wide character buffer going from UTF8 encoded chars -> UTF16 (windows
-  // wchar) this step is necessary when using MultiByteToWideChar
+  // wide character buffer going from UTF8 encoded chars -> UTF16 (windows wchar) 
+  // this step is necessary when using MultiByteToWideChar
   mSize = MultiByteToWideChar(CP_UTF8, 0, str.c_str(), -1, NULL, 0);
   wchar_t* wstr = (wchar_t*)zAllocate(mSize * sizeof(wchar_t));
-  // using the acquired information needed allocate a destination buffer and
-  // covert the utf8 encoded character string to a wide string
+  // using the acquired information needed allocate a destination buffer and covert 
+  // the utf8 encoded character string to a wide string
   MultiByteToWideChar(CP_UTF8, 0, str.c_str(), -1, wstr, mSize);
   mData.SetData((::byte*)wstr, mSize * sizeof(wchar_t), true);
 }
@@ -123,10 +122,10 @@ String Narrow(const wchar_t* wstr)
   // first read the number of wide characters we need to allocate for our
   // character buffer going from UTF16 (windows wchar) -> UTF8 encoded chars
   // this step is necessary when using WideCharToMultiByte
-  size_t size = WideCharToMultiByte(CP_UTF8, 0, wstr, -1, NULL, 0, NULL, NULL);
+  int size = WideCharToMultiByte(CP_UTF8, 0, wstr, -1, NULL, 0, NULL, NULL);
   char* str = new char[size];
-  // using the acquired information needed allocate a destination buffer and
-  // covert the wide string to a utf8 encoded character string
+  // using the acquired information needed allocate a destination buffer and covert 
+  // the wide string to a utf8 encoded character string
   WideCharToMultiByte(CP_UTF8, 0, wstr, -1, str, size, NULL, NULL);
   String ret(str);
   delete[] str;
@@ -146,12 +145,12 @@ WString Widen(const char* str)
   if (!str)
     return WString();
   // first read the number of wide characters we need to allocate for our
-  // wide character buffer going from UTF8 encoded chars -> UTF16 (windows
-  // wchar) this step is necessary when using MultiByteToWideChar
-  size_t size = MultiByteToWideChar(CP_UTF8, 0, str, -1, NULL, 0);
+  // wide character buffer going from UTF8 encoded chars -> UTF16 (windows wchar) 
+  // this step is necessary when using MultiByteToWideChar
+  int size = MultiByteToWideChar(CP_UTF8, 0, str, -1, NULL, 0);
   wchar_t* wstr = new wchar_t[size];
-  // using the acquired information needed allocate a destination buffer and
-  // covert the utf8 encoded character string to a wide string
+  // using the acquired information needed allocate a destination buffer and covert 
+  // the utf8 encoded character string to a wide string
   MultiByteToWideChar(CP_UTF8, 0, str, -1, wstr, size);
   WString ret(wstr);
   delete[] wstr;
@@ -172,7 +171,7 @@ int Utf16ToUtf8(int utf16)
   // take the straight code point value and put it in a wchar
   wchar_t inputUTF16 = (wchar_t)utf16;
 
-  size_t size = WideCharToMultiByte(CP_UTF8, 0, &inputUTF16, 1, NULL, 0, NULL, NULL);
+  int size = WideCharToMultiByte(CP_UTF8, 0, &inputUTF16, 1, NULL, 0, NULL, NULL);
 
   // output buffer is exact required size, no null terminator processed
   char outputUTF8[4] = {0};
@@ -180,7 +179,7 @@ int Utf16ToUtf8(int utf16)
 
   // after converting to UTF8 pack the code points back into an int
   int key = 0;
-  for (size_t i = 0; i < size; ++i)
+  for (int i = 0; i < size; ++i)
   {
     key <<= 8;
     key += (uchar)outputUTF8[i];
